@@ -18,7 +18,7 @@ func Test_runVersion(t *testing.T) {
 		{"1", args{&cobra.Command{}, []string{}}},
 	}
 	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
+		t.Run(tt.name, func(_ *testing.T) {
 			runVersion(tt.args.in0, tt.args.in1)
 		})
 	}
