@@ -34,8 +34,7 @@ func TestNewFormatter(t *testing.T) {
 	}
 }
 
-type mockWriter struct {
-}
+type mockWriter struct{}
 
 func (m *mockWriter) Write(p []byte) (n int, err error) {
 	return 0, nil
