@@ -2,6 +2,7 @@ package formatter
 
 import (
 	"bytes"
+	"fmt"
 	"log"
 	"os"
 )
@@ -10,7 +11,7 @@ type stdout struct {
 	l *log.Logger
 }
 
-func newStdoutFormatter(logger *log.Logger) Formatter {
+func newStdoutFormatter(logger *log.Logger) *stdout {
 	return &stdout{
 		l: logger,
 	}
@@ -20,7 +21,7 @@ func (s *stdout) Output(b bytes.Buffer) error {
 	_, err := b.WriteTo(os.Stdout)
 	if err != nil {
 		s.l.Printf("error: cannot write to stdout: %s", err)
-		return err
+		return fmt.Errorf("cannot write to stdout: %w", err)
 	}
 	return nil
 }
